Create uncheckedcallreturnvalue.sol
# Unchecked Call return Value

In Solidity you can use a few low-level call methods that work on raw addresses: call, callcode, delegatecall, and send. 
These low-level methods never throw an exception but will return false if the call encounters an exception.

The bug exists where a send is used without checking the response. 
If a winner's transaction fails it allows payedOut to be set to true (regardless of whether ether was sent or not). 
In this case, the public can withdraw the winner's winnings using the withdrawLeftOver function.

# Mitigation
It is recommended to use the transfer function rather than send. This is because transfer will revert if the external transaction reverts.
Send or transfer functions may be affected by changes in the cost of operation in Ethereum, So use 
                    msg.sender.call{value: balance}("")

If you choose to use the low-level call methods, make sure to handle the possibility that the call will fail, by checking the return value.