integer underflow example
The require statement in “transfer” function can be bypassed using an underflow. 
Consider a user that has no balance. They could call the transfer() function with any non-zero _value and pass the require statement
 This is because balances[msg.sender] is zero (and a uint256) 
so subtracting any positive amount (excluding 2^256) will result in a positive number due to the underflow