defaultvisibility example
The default visibility for functions is public. 
Therefore functions that do not specify any visibility will be callable by external users. 
The issue comes when developers mistakenly ignore visibility specifiers on functions which should be private (or only callable within the contract itself).

This simple contract is designed to act as an address guessing bounty game. 
To win the balance of the contract, a user must generate an Ethereum address whose last 8 hex characters are 0. 
Once obtained, they can call the WithdrawWinnings() function to obtain their bounty.

Unfortunately, the visibility of the functions have not been specified.
 In particular, the _sendWinnings() function is public and thus any address can call this function to steal the bounty.