delegate.sol real life example
DELEGATECALL opcode is identical to the standard message call, except that the code executed at the targeted address is run in the context of the calling contract along with the fact that msg.sender and msg.value remain unchanged. 
This feature enables the implementation of libraries whereby developers can create reusable code for future contracts.
WalletLibrary contract is itself a contract and maintains it's own state.
It is possible to send calls to the WalletLibrary contract itself. Specifically, the WalletLibrary contract could be initialised, and become owned. 
A user did this, by calling initWallet() function on the WalletLibrary contract, becoming an owner of the library contract. The same user, subsequently called the kill() function. 
Because the user was an owner of the Library contract, the modifier passed and the library contract suicided. As all Wallet contracts in existence refer to this library contract and contain no method to change this reference, all of their functionality, including the ability to withdraw ether is lost along with the WalletLibrary contract. 
More directly, all ether in all parity multi-sig wallets of this type instantly become lost or permanently unrecoverable.


# Preventive measures
Solidity provides the library keyword for implementing library contracts 
 This ensures the library contract is stateless and non-self-destructable.
when using DELEGATECALL pay careful attention to the possible calling context of both the library contract and the calling contract, and whenever possible, build state-less libraries.