blocktimestampmanipulation example
The roulette contract  is a game where you can win all of the Ether in the contract if you can submit a transaction at a specific time. 
The contract contains 10 Ether. To play the game you call the function spin and send the contract 1 Ether. 
If the block.timestamp is divisible by 7 you win all the Ether in the contract.
Attacker
Call the spin function and submit 1 Ether to play the game
Then submit a block timestamp for the next block that is divisible by 7


# PRNG based random generators
PRNGs based on block variables
There are a number of block variables that may be wrongly used as a source of entropy:

block.coinbase represents the address of the miner who mined the current block.
block.difficulty is a relative measure of how difficult it was to find the block.
block.gaslimit restricts maximum gas consumption for transactions within the block.
block.number is the height of current block.
block.timestamp is when the block was mined.

All these block variables can be manipulated by miners, so they cannot be used as a source of entropy because of the miners’ incentive. 
The block variables are obviously shared within the same block. So if an attacker’s contract calls the victim contract via an internal message, the same PRNG in both contracts will yield the same outcome.