denial of service example 2
when a contract may iterate through an array to pay users (e.g., supporters in a crowdfunding contract). 
It's common to want to make sure that each payment succeeds. 
If not, one should revert. 
The issue is that if one call fails, you are reverting the whole payout system, meaning the loop will never complete. 
No one gets paid because one address is forcing an error.