denial of service example 2
when a contract may iterate through an array to pay users (e.g., supporters in a crowdfunding contract). 
It's common to want to make sure that each payment succeeds. 
If not, one should revert. 
The issue is that if one call fails, you are reverting the whole payout system, meaning the loop will never complete. 
No one gets paid because one address is forcing an error.

By paying out to everyone at once, you risk running into the block gas limit.
This can lead to problems even in the absence of an intentional attack. 
it's especially bad if an attacker can manipulate the amount of gas needed. 
In the case of the previous example, the attacker could add a bunch of addresses, each of which needs to get a very small refund. 
The gas cost of refunding each of the attacker's addresses could, therefore, end up being more than the gas limit, blocking the refund transaction from happening at all.